backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.7
starlette==0.47.2
python-dotenv==1.1.1
httpx==0.27.0
python-multipart==0.0.7